+++ src/lib/emoji/EmojiPreview.svelte
@@ -44,18 +44,18 @@
     </span>
     <div class="tip-container" aria-hidden="true">
       <span class="tip-navigate">
-        <key class="icon-upvote"></key> &amp;
-        <key class="icon-downvote"></key> ({t(
+        <kbd class="icon-upvote"></kbd> &amp;
+        <kbd class="icon-downvote"></kbd> ({t(
           'autocomplete.preview.navigate',
         )})</span
       >
       <span class="tip-complete">
-        <key>TAB</key> or <key>ENTER</key> ({t(
+        <kbd>TAB</kbd> or <kbd>ENTER</kbd> ({t(
           'autocomplete.preview.select',
         )})</span
       >
       <span class="tip-close"
-        ><key>ESC</key> ({t('autocomplete.preview.close')})</span
+        ><kbd>ESC</kbd> ({t('autocomplete.preview.close')})</span
       >
     </div>
   </div>