+++ src/lib/satellites/Eta.svelte
@@ -3,7 +3,7 @@
   import {
     getQueuePosition,
     getCurrentSongMinutes,
-    PLAYER_SHARING_CONTAINER,
+    PLAYER_BUTTONS_CONTAINER,
   } from '../queup.ui';
   import { t } from '../stores/i18n.svelte';
 
@@ -40,14 +40,29 @@
 </script>
 
 <button
-  use:teleport={{ to: PLAYER_SHARING_CONTAINER }}
+  use:teleport={{ to: PLAYER_BUTTONS_CONTAINER }}
   id="dubplus-eta"
   aria-label={eta}
   type="button"
-  class="icon-history eta_tooltip_t dubplus-btn-player"
+  class="text-white/50 hover:text-white transition-colors eta_tooltip_t"
   data-dp-tooltip={eta}
   onmouseenter={() => {
     eta = getEta();
   }}
 >
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="24"
+    height="24"
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke="currentColor"
+    stroke-width="2"
+    stroke-linecap="round"
+    stroke-linejoin="round"
+    class="lucide lucide-rotate-ccw-clock-icon lucide-rotate-ccw-clock"
+    ><path d="M3 12a9 9 0 1 0 9-9 9.75 9.75 0 0 0-6.74 2.74L3 8" /><path
+      d="M3 3v5h5"
+    /><path d="M12 7v5l4 2" /></svg
+  >
 </button>